// apps/backend/src/app/modules/blockchain/types/api.ts
export interface SignerPayloadJSONDto {
  /**
   * The ss-58 encoded address
   * @example yGCyN3eydMkze4EPtz59Tn7obwbUbYNZCz48dp8FRdemTaLwm
   */
  address: string;

  /** The checkpoint hash of the block, in hex */
  blockHash: string;

  /** The checkpoint block number, in hex */
  blockNumber: string;

  /** The era for this transaction, in hex */
  era: string;

  /** The genesis hash of the chain, in hex */
  genesisHash: string;

  /** The encoded method (with arguments) in hex */
  method: string;

  /** The nonce for this transaction, in hex */
  nonce: string;

  /** The current spec version for the runtime */
  specVersion: string;

  /** The tip for this transaction, in hex */
  tip: string;

  /** The current transaction version for the runtime */
  transactionVersion: string;

  /** The applicable signed extensions for this runtime */
  signedExtensions: string[];

  /** The version of the extrinsic we are dealing with */
  version: number;
}

export interface SignerPayloadRawDto {
  /**
   * The ss-58 encoded address
   * @example yGCyN3eydMkze4EPtz59Tn7obwbUbYNZCz48dp8FRdemTaLwm
   */
  address: string;

  /** The hex-encoded data for this request */
  data: string;

  /** The type of the contained data */
  type: 'bytes' | 'payload';
}

export interface FeeResponse {
  /** @example 92485000000000000 */
  raw: string;

  /** @example 0.092485000000000000 */
  amount: string;

  /** @example 92.4850 m */
  formatted: string;

  /** @example UNQ */
  unit: string;

  /** @example 18 */
  decimals: number;
}

export interface UnsignedTxPayloadResponse {
  signerPayloadJSON: SignerPayloadJSONDto;
  signerPayloadRaw: SignerPayloadRawDto;
  signerPayloadHex: string;
  fee?: FeeResponse;
}
// ...
export interface CollectionSponsorship {
  /**
   * The ss-58 encoded address
   * @example yGCyN3eydMkze4EPtz59Tn7obwbUbYNZCz48dp8FRdemTaLwm
   */
  address: string;

  /** @example false */
  isConfirmed: boolean;
}

export interface CollectionLimitsDto {
  /**
   * Maximum number of tokens that one address can own
   * @example 1000
   */
  accountTokenOwnershipLimit?: number;

  /**
   * Maximum byte size of custom token data that can be sponsored when tokens are minted in sponsored mode
   * @example 1024
   */
  sponsoredDataSize?: number;

  /**
   * Defines how many blocks need to pass between setVariableMetadata transactions in order for them to be sponsored
   * @example 30
   */
  sponsoredDataRateLimit?: number;

  /**
   * Total amount of tokens that can be minted in this collection
   * @example 1000000
   */
  tokenLimit?: number;

  /**
   * Time interval in blocks that defines once per how long a non-privileged user transfer or mint transaction can be sponsored
   * @example 6
   */
  sponsorTransferTimeout?: number;

  /**
   * Time interval in blocks that defines once per how long a non-privileged user approve transaction can be sponsored
   * @example 6
   */
  sponsorApproveTimeout?: number;

  /**
   * Boolean value that tells if collection owner or admins can transfer or burn tokens owned by other non-privileged users
   * @example false
   */
  ownerCanTransfer?: boolean;

  /**
   * Boolean value that tells if collection owner can destroy it
   * @example false
   */
  ownerCanDestroy?: boolean;

  /**
   * Flag that defines whether token transfers between users are currently enabled
   * @example false
   */
  transfersEnabled?: boolean;
}

export interface CollectionNestingPermissionsDto {
  tokenOwner: boolean;
  collectionAdmin: boolean;
}

export interface CollectionPermissionsDto {
  access?: 'Normal' | 'AllowList';
  mintMode?: boolean;
  nesting?: CollectionNestingPermissionsDto;
}
// ...
export interface SubmitResponse {
  hash: string;
  fee?: FeeResponse;
}
// ...
export interface AttributeSchemaDto {
  enumValues?: Record<string, string | number | Record<string, string>>;
  name: string | Record<string, string>;
  optional: boolean;
  kind: 'enum' | 'enumMultiple' | 'freeValue';
  type:
    | 'integer'
    | 'float'
    | 'boolean'
    | 'timestamp'
    | 'string'
    | 'url'
    | 'isoDate'
    | 'time'
    | 'colorRgba'
    | 'localizedStringDictionary';
}

export interface ImageDto {
  /** @example https://ipfs.unique.network/ipfs/{infix}.ext */
  urlTemplate: string;
}

export interface OldPropertiesDto {
  _old_schemaVersion: string;
  _old_offchainSchema: string;
  _old_constOnChainSchema: string;
  _old_variableOnChainSchema: string;
}

export interface ImagePreviewDto {
  /** @example https://ipfs.unique.network/ipfs/{infix}.ext */
  urlTemplate?: string;
}

export interface AudioDto {
  /** @example https://ipfs.unique.network/ipfs/{infix}.ext */
  urlTemplate?: string;
  format: string;
  isLossless: boolean;
}

export interface SpatialObjectDto {
  /** @example https://ipfs.unique.network/ipfs/{infix}.ext */
  urlTemplate?: string;
  format: string;
}

export interface VideoDto {
  /** @example https://ipfs.unique.network/ipfs/{infix}.ext */
  urlTemplate?: string;
}
// ...
export interface UniqueCollectionSchemaToCreateDto {
  /** @example {"0":{"name":{"en":"gender"},"type":"localizedStringDictionary","kind":"enum","enumValues":{"0":{"en":"Male"},"1":{"en":"Female"}}},"1":{"name":{"en":"traits"},"type":"localizedStringDictionary","kind":"enumMultiple","enumValues":{"0":{"en":"Black Lipstick"},"1":{"en":"Red Lipstick"}}}} */
  attributesSchema: Record<string, AttributeSchemaDto>;

  /** @example 1.0.0 */
  attributesSchemaVersion: string;
  coverPicture:
    | { urlInfix?: string; hash?: string | null }
    | { url?: string; hash?: string | null }
    | { ipfsCid?: string; hash?: string | null };
  image: ImageDto;

  /** @example unique */
  schemaName: 'unique' | '_old_';

  /** @example 1.0.0 */
  schemaVersion: string;
  coverPicturePreview:
    | { urlInfix?: string; hash?: string | null }
    | { url?: string; hash?: string | null }
    | { ipfsCid?: string; hash?: string | null };
  imagePreview: ImagePreviewDto;
  audio: AudioDto;
  spatialObject: SpatialObjectDto;
  video: VideoDto;
}

export interface CreateCollectionNewRequest {
  mode?: 'Nft' | 'Fungible' | 'ReFungible';

  /** @example Sample collection name */
  name: string;

  /** @example sample collection description */
  description: string;

  /** @example TEST */
  tokenPrefix: string;
  sponsorship?: CollectionSponsorship;
  limits?: CollectionLimitsDto;
  metaUpdatePermission?: 'ItemOwner' | 'Admin' | 'None';
  permissions?: CollectionPermissionsDto;

  /**
   * The ss-58 encoded address
   * @example yGCyN3eydMkze4EPtz59Tn7obwbUbYNZCz48dp8FRdemTaLwm
   */
  address: string;
  schema: Partial<UniqueCollectionSchemaToCreateDto>;
}
// ...
export interface SubmitTxBody {
  signerPayloadJSON: SignerPayloadJSONDto;

  /** Warning: Signature must be with SignatureType! */
  signature: string;
}

export interface VerificationResultResponse {
  isValid: boolean;
  errorMessage: string;
}

export interface SubmitResultResponse {
  hash: string;
}

export interface ExtrinsicResultEvent {
  section: string;
  method: string;
  data: object;
}

export interface ExtrinsicResultResponse {
  status: string;
  isCompleted: boolean;
  isError: boolean;
  blockHash: string;
  blockIndex: number;
  errorMessage: string;
  events: Array<ExtrinsicResultEvent>;
}